requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
googleapis-common-protos==1.55.0
grpcio==1.32.0
grpcio-tools==1.44.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.1.0
numpy==1.22.3
pandas==1.4.1
protobuf==3.19.4
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
urllib3==1.26.8
Werkzeug==2.0.3
zemberek-grpc==0.16.1
